Change e2e tests to accommodate fuse oddities (see msg)
On a write after fsyncing the OS (or FUSE)
still returns EOF with 0 bytes, even when spork itself returned
the correct file contents and the correct file attributes (updated
size and updated modified time). For now reading from the same
file handle will not return the written content. The handle needs to be
closed and the file reopened.